# openapi.yaml
openapi: 3.0.0
info:
  title: Item Service API
  version: v1
  description: API for managing items.

paths:
  /items:
    get:
      summary: List all items
      operationId: getItems
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new item
      operationId: createItem
      requestBody:
        description: Item to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewItem'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /items/{id}:
    get:
      summary: Get an item by its ID
      operationId: getItemById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the item to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update an existing item
      operationId: updateItemById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the item to update
          schema:
            type: integer
            format: int64
      requestBody:
        description: Item data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItem'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid request payload or input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an item by its ID
      operationId: deleteItemById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the item to delete
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Item deleted successfully
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    NewItem:
      type: object
      required:
        - name
        - priority
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: integer
          format: int32
    UpdateItem:
      type: object
      required:
        - name
        - priority
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: integer
          format: int32
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: integer
          format: int32
      required:
        - name
        - priority
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error